Move to another namespace and class name
